feature: introduce parallel book scraping
